+++ init.lua
@@ -82,7 +82,6 @@ include('SpatialConvolution.lua')
 include('SpatialFullConvolution.lua')
 include('SpatialFullConvolutionMap.lua')
 include('SpatialConvolutionMM.lua')
-include('SpatialDeconvolution.lua')
 include('SpatialConvolutionMap.lua')
 include('SpatialSubSampling.lua')
 include('SpatialMaxPooling.lua')
@@ -101,7 +100,7 @@ include('SpatialUpSamplingNearest.lua')
 include('SpatialBatchNormalization.lua')
 
 include('VolumetricConvolution.lua')
-include('VolumetricDeconvolution.lua')
+include('VolumetricFullConvolution.lua')
 include('VolumetricMaxPooling.lua')
 include('VolumetricAveragePooling.lua')
